+++ platformio.ini
@@ -142,7 +142,7 @@ extra_scripts =
 lib_deps = 
 	ArduinoJson@>=7.0.0
   elims/PsychicMqttClient@^0.2.4
-  ElectronicCats/MPU6050 @ 1.3.0 ; for D_MPU6050 driver
+  ElectronicCats/MPU6050 @ 1.3.0 ; for D_IMU.h driver
 
 ;💫
 [moonlight]
